Separate FilterBar component
Filtering itself does not work yet, but the component looks and behaves
correctly.
