+++ frontend/src/components/CustomTable.tsx
@@ -15,19 +15,11 @@
  */
 
 import * as React from 'react';
-import * as Autosuggest from 'react-autosuggest';
-// TODO match and parse required editing the node modules to import correctly.
-// Perhaps we should consider using allowSyntheticDefaultImports
-// tslint:disable-next-line:no-var-requires
-const match = require('autosuggest-highlight/match');
-// tslint:disable-next-line:no-var-requires
-const parse = require('autosuggest-highlight/parse');
 import ArrowRight from '@material-ui/icons/ArrowRight';
 import Checkbox, { CheckboxProps } from '@material-ui/core/Checkbox';
 import ChevronLeft from '@material-ui/icons/ChevronLeft';
 import ChevronRight from '@material-ui/icons/ChevronRight';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import ChipInput, { ChipRendererArgs } from 'material-ui-chip-input';
 import IconButton from '@material-ui/core/IconButton';
 import MenuItem from '@material-ui/core/MenuItem';
 import Radio from '@material-ui/core/Radio';
@@ -40,9 +32,8 @@ import { ListRequest } from '../lib/Apis';
 import { classes, stylesheet } from 'typestyle';
 import { fonts, fontsize, dimension, commonCss, color, padding } from '../Css';
 import { logger } from '../lib/Utils';
-import Chip from '@material-ui/core/Chip';
-import Paper from '@material-ui/core/Paper';
-import { SuggestionsFetchRequestedParams, SuggestionHighlightedParams, SuggestionSelectedEventData } from 'react-autosuggest';
+import FilterBar from './FilterBar';
+import { ApiFilter } from 'src/apis/filter';
 
 export enum ExpandState {
   COLLAPSED,
@@ -167,21 +158,12 @@ export const css = stylesheet({
   },
 });
 
-interface FilterChip {
-  key: string;
-  value: {
-    filterValue: number | string;
-    type: string;
-  };
-}
-
 interface CustomTableProps {
   columns: Column[];
   disablePaging?: boolean;
   disableSelection?: boolean;
   disableSorting?: boolean;
   emptyMessage?: string;
-  filterString?: string;
   getExpandComponent?: (index: number) => React.ReactNode;
   initialSortColumn?: string;
   initialSortOrder?: 'asc' | 'desc';
@@ -196,24 +178,16 @@ interface CustomTableProps {
 interface CustomTableState {
   currentPage: number;
   filterBy: string;
-  filterChips: FilterChip[];
-  filterShouldShowTypes: boolean;
   isBusy: boolean;
-  lastSelectedFilterType: string;
   maxPageIndex: number;
   pageSize: number;
   sortBy: string;
   sortOrder: 'asc' | 'desc';
-  suggestions: string[];
-  textFieldInput: string;
   tokenList: string[];
 }
 
 export default class CustomTable extends React.Component<CustomTableProps, CustomTableState> {
   private _isMounted = true;
-
-  private FILTER_TYPE_SUGGESTIONS = ['Name', 'Created at'];
-  private FILTER_VALUE_SUGGESTIONS = ['My Pipeline', 'Pipeline 1', 'Pipeline 2'];
 
   constructor(props: CustomTableProps) {
     super(props);
@@ -221,18 +195,12 @@ export default class CustomTable extends React.Component<CustomTableProps, Custo
     this.state = {
       currentPage: 0,
       filterBy: '',
-      filterChips: [],
-      // This will control whether we are suggesting the filter types (name, timestamp, etc.) or actual values
-      filterShouldShowTypes: true,
       isBusy: false,
-      lastSelectedFilterType: '',
       maxPageIndex: Number.MAX_SAFE_INTEGER,
       pageSize: 10,
       sortBy: props.initialSortColumn ||
         (props.columns.length ? props.columns[0].sortKey || '' : ''),
       sortOrder: props.initialSortOrder || 'desc',
-      suggestions: this.FILTER_TYPE_SUGGESTIONS,
-      textFieldInput: '',
       tokenList: [''],
     };
   }
@@ -284,9 +252,7 @@ export default class CustomTable extends React.Component<CustomTableProps, Custo
   }
 
   public render(): JSX.Element {
-    const { suggestions, pageSize, sortBy, sortOrder } = this.state;
-    // tslint:disable-next-line:no-console
-    console.log(suggestions);
+    const { pageSize, sortBy, sortOrder } = this.state;
     const numSelected = (this.props.selectedIds || []).length;
     const totalFlex = this.props.columns.reduce((total, c) => total += (c.flex || 1), 0);
     const widths = this.props.columns.map(c => (c.flex || 1) / totalFlex * 100);
@@ -295,40 +261,8 @@ export default class CustomTable extends React.Component<CustomTableProps, Custo
       <div className={commonCss.pageOverflowHidden}>
 
         {/* Filter/Search bar */}
-        <div>
-          <Autosuggest
-            getSuggestionValue={(suggestion) => suggestion}
-            inputProps={{
-              cancelBubble: true,
-              chips: this.state.filterChips,
-              classes,
-              onAdd: this._handleAddChip.bind(this),
-              onChange: this._handleTextFieldInputChange.bind(this),
-              onDelete: this._handleDeleteChip.bind(this),
-              value: this.state.textFieldInput,
-            }}
-            onSuggestionHighlighted={this._handleSuggestionHighlighted.bind(this)}
-            onSuggestionSelected={this._onSuggestionSelected.bind(this)}
-            onSuggestionsClearRequested={this._handleSuggestionsClearRequested.bind(this)}
-            onSuggestionsFetchRequested={this._handleSuggestionsFetchRequested.bind(this)}
-            renderInputComponent={this._renderChipInput.bind(this)}
-            renderSuggestion={this._renderSuggestion.bind(this)}
-            renderSuggestionsContainer={this._renderSuggestionsContainer.bind(this)}
-            shouldRenderSuggestions={this._shouldRenderSuggestion.bind(this)}
-            suggestions={suggestions}
-          />
-        </div>
-        <br />
-        <br />
-        <div>Current value of "textFieldInput":</div>
-        <div style={{fontFamily: 'monospace'}}>{this.state.textFieldInput}</div>
-        <br />
-        <div>filterShouldShowTypes:</div>
-        <div style={{fontFamily: 'monospace'}}>{'' + this.state.filterShouldShowTypes}</div>
-        <br />
-        <div>lastSelectedFilterType:</div>
-        <div style={{fontFamily: 'monospace'}}>{'' + this.state.lastSelectedFilterType}</div>
-
+        {/* TODO: Do not hardcode these, they should be CustomTable props passed down. */}
+        <FilterBar filter={this._requestFilter} filterTypes={['Name', 'Created at']} />
 
         {/* Header */}
         <div className={classes(
@@ -486,234 +420,12 @@ export default class CustomTable extends React.Component<CustomTableProps, Custo
     return result;
   }
 
-  private _renderChipInput(inputProps: any): JSX.Element {
-    // TODO we need to pull 'classes' out of the props, but it conflicts with the 'classes' from type style
-    // tslint:disable-next-line:no-shadowed-variable
-    const { classes, value, cancelBubble, onChange, onAdd, onDelete, ref, ...other } = inputProps;
-
-    return (
-    <ChipInput
-      classes={{}}
-      InputProps={{
-        // value here is the key to updating what is displayed in the bar.
-        value,
-      }}
-      value={this.state.filterChips}
-      onUpdateInput={onChange}
-      onAdd={onAdd}
-      onDelete={onDelete}
-      // this ref appears to be an empty function, but seems to be necessary
-      inputRef={ref}
-      {...other}
-      chipRenderer={(
-        args: ChipRendererArgs,
-        key: any,
-      ) => {
-        return (
-        <Chip
-          key={key}
-          style={{
-            backgroundColor: args.isFocused ? 'cyan' : 'lightgray',
-            pointerEvents: args.isDisabled ? 'none' : undefined,
-          }}
-          onClick={args.handleClick}
-          onDelete={args.handleDelete}
-          label={(
-            <div>
-              {(args.value as any).value.type &&
-                <span style={{ color: 'gray' }}>{(args.value as any).value.type} : </span>
-              }
-              <span>{(args.value as any).value.filterValue}</span>
-            </div>
-          )}
-        />
-      );}}
-    />);
-  }
-
-  private _renderSuggestion (suggestion: string, obj: { query: string, isHighlighted: boolean }): JSX.Element {
-    // TODO: this matching is fuzzy, but the rest of the code in this file uses 'startsWith'
-    const matches = match(suggestion, obj.query);
-    const parts = parse(suggestion, matches);
-    // tslint:disable-next-line:no-console
-    console.log(suggestion, obj.query, matches, parts);
-
-    return (
-      <MenuItem
-        selected={obj.isHighlighted}
-        component='div'
-        onMouseDown={(e) => e.preventDefault()} // prevent the click causing the input to be blurred
-      >
-        <div>
-          {parts.map((part: any, index: any) => {
-            return part.highlight ? (
-              <span key={String(index)} style={{ fontWeight: 600 }}>
-                {part.text}
-              </span>
-            ) : (
-              <strong key={String(index)} style={{ fontWeight: 300 }}>
-                {part.text}
-              </strong>
-            );
-          })}
-        </div>
-      </MenuItem>
+  private _requestFilter = async (filter: ApiFilter) => {
+    // TODO: verify this works as intended.
+    this._resetToFirstPage(
+      // Stringify the filter, and convert it to base64
+      await this.reload({ filterBy: btoa(JSON.stringify(filter)) })
     );
-  }
-
-  private async _handleSuggestionsFetchRequested(param: SuggestionsFetchRequestedParams): Promise<void> {
-    // tslint:disable-next-line:no-console
-    console.log('handleSuggestionsFetchRequested', param);
-    // TODO: actually fetch suggestions
-    const possibleSuggestions =
-      this.state.filterShouldShowTypes
-      ? this.FILTER_TYPE_SUGGESTIONS
-      : this.FILTER_VALUE_SUGGESTIONS;
-    this.setStateSafe({
-      suggestions: possibleSuggestions.filter(s => {
-        const trimmedInput = param.value.substring(this.state.lastSelectedFilterType.length);
-        // tslint:disable-next-line:no-console
-        // console.log(s.toLocaleLowerCase().startsWith(trimmedInput.toLocaleLowerCase()));
-        return s.toLocaleLowerCase().startsWith(trimmedInput.toLocaleLowerCase());
-      })
-    });
-  }
-
-  private async _handleSuggestionsClearRequested(): Promise<void> {
-    // TODO: anything else need to happen?
-    // tslint:disable-next-line:no-console
-    console.log('clear! (actually reseting, not clearing)');
-    const possibleSuggestions =
-      this.state.filterShouldShowTypes
-        ? this.FILTER_TYPE_SUGGESTIONS
-        : this.FILTER_VALUE_SUGGESTIONS;
-    this.setStateSafe({ suggestions: possibleSuggestions });
-  }
-
-  private _renderSuggestionsContainer (options: any): JSX.Element {
-    const { containerProps, children } = options;
-    return (
-      <Paper {...containerProps} square={true} style={{ width: 120 }} >
-        {children}
-      </Paper>
-    );
-  }
-
-  private _handleSuggestionHighlighted(params: SuggestionHighlightedParams): void {
-    // TODO: this will need to be updated if suggestion becomes more than a string.
-    // tslint:disable-next-line:no-console
-    console.log('handleSuggestionHighlighted', this.state.filterShouldShowTypes, params);
-    if (params && params.suggestion) {
-      this.setStateSafe({
-        textFieldInput: this.state.filterShouldShowTypes
-          ? params.suggestion
-          : this.state.lastSelectedFilterType + params.suggestion
-      });
-    }
-  }
-
-  private _handleTextFieldInputChange = (event: React.FormEvent<any>, param: Autosuggest.ChangeEvent) => {
-    // tslint:disable-next-line:no-console
-    console.log('handletextFieldInputChange', event, param);
-    const { lastSelectedFilterType } = this.state;
-    if (param.method === 'type') {
-      const newValLower = param.newValue.toLocaleLowerCase();
-      // Check if the current text in the input field matches any of our filter types, and ensure there is a colon.
-      const inputFilterTypePrefix = this.FILTER_TYPE_SUGGESTIONS.find((s) => newValLower.startsWith(s.toLocaleLowerCase() + ':'));
-      this.setStateSafe({
-        // Should show types if the input does not start with a filter type.
-        filterShouldShowTypes: !inputFilterTypePrefix,
-        // lastSelectedFilterType should be cleared if input no longer starts with it.
-        lastSelectedFilterType: inputFilterTypePrefix ? inputFilterTypePrefix + ':' : '',
-        textFieldInput: param.newValue,
-      }, async () => this._handleSuggestionsFetchRequested({ value: param.newValue, reason: 'input-changed' }) );
-    } else if (param.method === 'up' || param.method === 'down') {
-      this.setStateSafe({
-        textFieldInput: lastSelectedFilterType + param.newValue
-      });
-    }
-  };
-
-  private _shouldRenderSuggestion(value: string): boolean {
-    const { lastSelectedFilterType, suggestions } = this.state;
-    // tslint:disable-next-line:no-console
-    console.log('shouldRenderSuggestions', lastSelectedFilterType, value, suggestions);
-    return !value
-      || !!suggestions.find(s => {
-        const trimmedInput = value.substring(lastSelectedFilterType.length);
-        // tslint:disable-next-line:no-console
-        // console.log(s.toLocaleLowerCase().startsWith(trimmedInput.toLocaleLowerCase()));
-        return s.toLocaleLowerCase().startsWith(trimmedInput.toLocaleLowerCase());
-      });
-  }
-
-  private _handleAddChip(chip: string): void {
-    // tslint:disable-next-line:no-console
-    console.log('_handleAddChip', chip);
-    const { filterChips, lastSelectedFilterType } = this.state;
-    filterChips.push({
-      key: filterChips.length + '',
-      value: {
-        filterValue: `${chip}`,
-        // Trim off the ":" at the end of the filter type
-        type: lastSelectedFilterType.slice(0, -1),
-      }
-    });
-    this.setStateSafe({
-        filterChips,
-        filterShouldShowTypes: true,
-        // Clear the last selected suggestion so the autosuggest will no longer look for it.
-        lastSelectedFilterType: '',
-        textFieldInput: ''
-      },
-      async () => this._requestFilter()
-    );
-  }
-
-  private _handleDeleteChip(chip: FilterChip, index: number): void {
-    // tslint:disable-next-line:no-console
-    // console.log('_handleDeleteChip', chip, index);
-    const chips = this.state.filterChips;
-    // tslint:disable-next-line:no-console
-    console.log('_handleDeleteChip', index, chips);
-    chips.splice(index, 1);
-    this.setStateSafe({ filterChips: chips }, async () => this._requestFilter());
-  }
-
-
-  private async _requestFilter(): Promise<void> {
-    if (this.state.filterChips.length) {
-      this._resetToFirstPage(
-        // TODO: convert this to actual filter, not just JSON string.
-        await this.reload({ filterBy: JSON.stringify(this.state.filterChips) })
-      );
-    }
-  }
-
-  private _onSuggestionSelected(event: React.FormEvent<any>, data: SuggestionSelectedEventData<string>): void {
-    // tslint:disable-next-line:no-console
-    console.log('onSuggestionSelected');
-
-    const { filterShouldShowTypes } = this.state;
-    let { lastSelectedFilterType } = this.state;
-    let textFieldInput;
-
-    // Don't add a chip if we only selected a type.
-    if (filterShouldShowTypes) {
-      // Hold onto the last type that was selected so it can be subtracted from the query, since it
-      // will display as text in the input box.
-      lastSelectedFilterType = data.suggestion + ':';
-      textFieldInput = lastSelectedFilterType;
-      this.setStateSafe({
-        filterShouldShowTypes: !filterShouldShowTypes,
-        lastSelectedFilterType,
-        textFieldInput,
-      });
-    } else {
-      // _handleAddChip covers clearing the text field and such so we don't do it here.
-      this._handleAddChip(data.suggestion);
-    }
-    return event.preventDefault();
   }
 
   private _requestSort(sortBy?: string): void {
